Replace GdkRectangle in PresentationInterface api with Rectangle<double>

+++ inc/scroom/cairo-helpers.hh
@@ -25,12 +25,12 @@ void setClip(cairo_t* cr, int x, int y, int width, int height);
 void setClip(cairo_t* cr, const GdkRectangle& area);
 void drawRectangle(cairo_t* cr, Color const& c, Rectangle<double> const& viewArea);
 void drawOutOfBoundsWithBackground(cairo_t* cr,
-                                   Rectangle<int> const& requestedPresentationArea,
-                                   Rectangle<int> const& actualPresentationArea, double pixelSize);
+                                   Rectangle<double> const& requestedPresentationArea,
+                                   Rectangle<double> const& actualPresentationArea, double pixelSize);
 void drawOutOfBoundsWithBackgroundColor(cairo_t* cr, const Color& background,
-                                   Rectangle<int> const& requestedPresentationArea,
-                                   Rectangle<int> const& actualPresentationArea, double pixelSize);
+                                   Rectangle<double> const& requestedPresentationArea,
+                                   Rectangle<double> const& actualPresentationArea, double pixelSize);
 void drawOutOfBoundsWithoutBackground(cairo_t* cr,
-                                      Rectangle<int> const& requestedPresentationArea,
-                                      Rectangle<int> const& actualPresentationArea, double pixelSize);
+                                      Rectangle<double> const& requestedPresentationArea,
+                                      Rectangle<double> const& actualPresentationArea, double pixelSize);
 double pixelSizeFromZoom(int zoom);
